add show array tutorial
